Fix update function to alter version in database
This makes it match the EE developer guide (http://ellislab.com/expressionengine/user-guide/development/extensions.html#activating-and-updating)
It also makes the devot:ee monitor detect the version update.  Otherwise, the version is not changed in the database.  EE will read the version number out of the file, but devot:ee will not so this makes the DB match the file.