Update ft.entry_type.php
This seems like bad news bears. You're creating a new instance of the Api_channel_fields which other add-ons may be relying on. I would either check to see if it already exists, and use it if if does, or create your own local reference and not mess with the global singleton. I believe this is what is causing conflicts with Publisher. When editing a field with Publisher installed it throws undefined property errors, e.g. "field_name" is not present.