Extra CSS to hide fields like Matrix and Playa
Related to https://github.com/rsanchez/entry_type/issues/54

I manage a site that was recently upgraded to EE 4.3.2. We noticed that entry type was not working completely for Matrix and Playa fields. The content was hidden, but there would be an overlaid bounding box where those fields would be in the layout if they were shown.

I've added some extra styling rules to the .entry-type-hidden class to hide these boxes completely